requirements.txt
alembic==1.6.5
aniso8601==9.0.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.20.46
click==8.0.0
Django==4.0.1
django-environ==0.8.1
django-storages==1.12.3
gunicorn==20.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
openpyxl==3.0.9
psycopg2==2.9.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
whitenoise==5.3.0
zipfile36==0.1.3

